+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -230,7 +230,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01K80QWCKMRCN081G75SDVRVFJ"
+        "01K80X5093N02KFDXEFDRZAM8P"
     )
     r"""The id of the resource"""