+++ packages/orq-rc/src/orq_ai_sdk/models/filelistop.py
@@ -79,7 +79,7 @@ class FileListData(BaseModel):
     workspace_id: str
     r"""The id of the resource"""
 
-    created: Optional[datetime] = parse_datetime("2025-10-20T12:00:33.324Z")
+    created: Optional[datetime] = parse_datetime("2025-10-20T13:32:37.564Z")
     r"""The date and time the resource was created"""