+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -132,7 +132,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01K80QWCKK6FT90K9XS9YY9Z55"
+        "01K80X5091SW48T2AGA68X77YQ"
     )
     r"""The id of the resource"""